# Optimize and benchmark handling of invalid keys