--- a/lightstep/sdk/metric/benchmark_test.go
+++ b/lightstep/sdk/metric/benchmark_test.go
@@ -19,8 +19,24 @@ import (
 	"testing"
 
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/data"
+	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/view"
 	"go.opentelemetry.io/otel/attribute"
 )
+
+// Tested prior to 0.11.0 release
+// goos: darwin
+// goarch: arm64
+// pkg: github.com/lightstep/otel-launcher-go/lightstep/sdk/metric
+// BenchmarkCounterAddNoAttrs-10                 	35354023	        33.79 ns/op	       0 B/op	       0 allocs/op
+// BenchmarkCounterAddOneAttr-10                 	14354538	        82.77 ns/op	      64 B/op	       1 allocs/op
+// BenchmarkCounterAddOneInvalidAttr-10          	 9307794	       128.4 ns/op	     128 B/op	       1 allocs/op
+// BenchmarkCounterAddManyAttrs-10               	 1000000	      1075 ns/op	     569 B/op	       6 allocs/op
+// BenchmarkCounterAddManyInvalidAttrs-10        	  832549	      1654 ns/op	    1080 B/op	      10 allocs/op
+// BenchmarkCounterAddManyFilteredAttrs-10       	 1000000	      1304 ns/op	     953 B/op	       8 allocs/op
+// BenchmarkCounterCollectOneAttrNoReuse-10      	 2537348	       468.0 ns/op	     400 B/op	       7 allocs/op
+// BenchmarkCounterCollectOneAttrWithReuse-10    	 3679694	       328.2 ns/op	     136 B/op	       3 allocs/op
+// BenchmarkCounterCollectTenAttrs-10            	  715490	      1635 ns/op	     712 B/op	      12 allocs/op
+// BenchmarkCounterCollectTenAttrsTenTimes-10    	   72478	     16475 ns/op	    7128 B/op	     120 allocs/op
 
 func BenchmarkCounterAddNoAttrs(b *testing.B) {
 	ctx := context.Background()
@@ -35,10 +51,6 @@ func BenchmarkCounterAddNoAttrs(b *testing.B) {
 	}
 }
 
-// Benchmark prints 3 allocs per Add():
-//  1. new []attribute.KeyValue for the list of attributes
-//  2. interface{} wrapper around attribute.Set
-//  3. an attribute array (map key)
 func BenchmarkCounterAddOneAttr(b *testing.B) {
 	ctx := context.Background()
 	rdr := NewManualReader("bench")
@@ -52,17 +64,19 @@ func BenchmarkCounterAddOneAttr(b *testing.B) {
 	}
 }
 
-// Benchmark prints 11 allocs per Add(), I see 10 in the profile:
-//  1. new []attribute.KeyValue for the list of attributes
-//  2. an attribute.Sortable (acquireRecord)
-//  3. the attribute.Set underlying array
-//  4. interface{} wrapper around attribute.Set value
-//  5. internal to sync.Map
-//  6. internal sync.Map
-//  7. new syncstate.record
-//  8. new viewstate.syncAccumulator
-//  9. an attribute.Sortable (findOutput)
-// 10. an output Aggregator
+func BenchmarkCounterAddOneInvalidAttr(b *testing.B) {
+	ctx := context.Background()
+	rdr := NewManualReader("bench")
+	provider := NewMeterProvider(WithReader(rdr))
+	b.ReportAllocs()
+
+	cntr, _ := provider.Meter("test").SyncInt64().Counter("hello")
+
+	for i := 0; i < b.N; i++ {
+		cntr.Add(ctx, 1, attribute.String("", "V"), attribute.String("K", "V"))
+	}
+}
+
 func BenchmarkCounterAddManyAttrs(b *testing.B) {
 	ctx := context.Background()
 	rdr := NewManualReader("bench")
@@ -73,6 +87,36 @@ func BenchmarkCounterAddManyAttrs(b *testing.B) {
 
 	for i := 0; i < b.N; i++ {
 		cntr.Add(ctx, 1, attribute.Int("K", i))
+	}
+}
+
+func BenchmarkCounterAddManyInvalidAttrs(b *testing.B) {
+	ctx := context.Background()
+	rdr := NewManualReader("bench")
+	provider := NewMeterProvider(WithReader(rdr))
+	b.ReportAllocs()
+
+	cntr, _ := provider.Meter("test").SyncInt64().Counter("hello")
+
+	for i := 0; i < b.N; i++ {
+		cntr.Add(ctx, 1, attribute.Int("", i), attribute.Int("K", i))
+	}
+}
+
+func BenchmarkCounterAddManyFilteredAttrs(b *testing.B) {
+	ctx := context.Background()
+	rdr := NewManualReader("bench")
+	provider := NewMeterProvider(
+		WithReader(rdr,
+			view.WithClause(view.WithKeys([]attribute.Key{"K"})),
+		),
+	)
+	b.ReportAllocs()
+
+	cntr, _ := provider.Meter("test").SyncInt64().Counter("hello")
+
+	for i := 0; i < b.N; i++ {
+		cntr.Add(ctx, 1, attribute.Int("L", i), attribute.Int("K", i))
 	}
 }