**Description:** Observed benchmarks had slipped when introducing handling of invalid keys in  #264 

**Link to tracking Issue:** n/a

**Testing:** New benchmarks added; unittests refined to be sure they cover all cases.

**Documentation:** n/a

BEFORE
```
goos: darwin
goarch: arm64
pkg: github.com/lightstep/otel-launcher-go/lightstep/sdk/metric
BenchmarkCounterAddNoAttrs-10                 	35070174	        34.01 ns/op	       0 B/op	       0 allocs/op
BenchmarkCounterAddOneAttr-10                 	14169888	        83.75 ns/op	      64 B/op	       1 allocs/op
BenchmarkCounterAddOneInvalidAttr-10          	 9203388	       129.6 ns/op	     128 B/op	       1 allocs/op
BenchmarkCounterAddManyAttrs-10               	 1000000	      1162 ns/op	     737 B/op	      10 allocs/op
BenchmarkCounterAddManyInvalidAttrs-10        	  877088	      1616 ns/op	    1230 B/op	      12 allocs/op
BenchmarkCounterAddManyFilteredAttrs-10       	 1000000	      1333 ns/op	     993 B/op	      10 allocs/op
BenchmarkCounterCollectOneAttrNoReuse-10      	 2551692	       468.2 ns/op	     400 B/op	       7 allocs/op
BenchmarkCounterCollectOneAttrWithReuse-10    	 3625554	       329.7 ns/op	     136 B/op	       3 allocs/op
BenchmarkCounterCollectTenAttrs-10            	  711798	      1639 ns/op	     712 B/op	      12 allocs/op
BenchmarkCounterCollectTenAttrsTenTimes-10    	   72469	     16401 ns/op	    7128 B/op	     120 allocs/op

```

AFTER
```
goos: darwin
goarch: arm64
pkg: github.com/lightstep/otel-launcher-go/lightstep/sdk/metric
BenchmarkCounterAddNoAttrs-10                 	34416027	        34.87 ns/op	       0 B/op	       0 allocs/op
BenchmarkCounterAddOneAttr-10                 	14471146	        82.80 ns/op	      64 B/op	       1 allocs/op
BenchmarkCounterAddOneInvalidAttr-10          	 9282039	       128.7 ns/op	     128 B/op	       1 allocs/op
BenchmarkCounterAddManyAttrs-10               	 1000000	      1033 ns/op	     569 B/op	       6 allocs/op
BenchmarkCounterAddManyInvalidAttrs-10        	  803008	      1680 ns/op	    1084 B/op	      10 allocs/op
BenchmarkCounterAddManyFilteredAttrs-10       	 1000000	      1346 ns/op	     954 B/op	       8 allocs/op
BenchmarkCounterCollectOneAttrNoReuse-10      	 2553843	       472.5 ns/op	     400 B/op	       7 allocs/op
BenchmarkCounterCollectOneAttrWithReuse-10    	 3628261	       328.6 ns/op	     136 B/op	       3 allocs/op
BenchmarkCounterCollectTenAttrs-10            	  720498	      1635 ns/op	     712 B/op	      12 allocs/op
BenchmarkCounterCollectTenAttrsTenTimes-10    	   73394	     16465 ns/op	    7128 B/op	     120 allocs/op
```